third step for combining split_dtype16 and split_stype32, split dtype 16 to two files
