+++ csrc/split_dtype.c
@@ -1,174 +1,13 @@
 #define PY_SSIZE_T_CLEAN
-#include "huf.h"
-#include "split_dtype_functions.h"
 #include <Python.h>
 #include <assert.h>
 #include <stdint.h>
 #include <time.h>
+#include "huf.h"
+#include "split_dtype_functions.h"
+#include "data_manipulation_dtype16.h" 
 
-///////////////////////////////////
-/// Split Helper Functions ///////
-//////////////////////////////////
-
-// Reordering function for float bits
-static uint32_t reorder_float_bits_dtype16(float number) {
-  union {
-    float f;
-    uint32_t u;
-  } value = {.f = number};
-
-  uint32_t sign = (value.u >> 8) & 0x800080;
-  uint32_t exponent = (value.u << 1) & 0xFF00FF00;
-  uint32_t mantissa = (value.u) & 0x7F007F;
-  return exponent | sign | mantissa;
-}
-
-// Helper function to reorder all floats in a bytearray
-static void reorder_all_floats_dtype16(u_int8_t *src, Py_ssize_t len) {
-  uint32_t *uint_array = (uint32_t *)src;
-  Py_ssize_t num_floats = len / sizeof(uint32_t);
-  for (Py_ssize_t i = 0; i < num_floats; i++) {
-    uint_array[i] = reorder_float_bits_dtype16(*(float *)&uint_array[i]);
-  }
-}
-
-// Helper function to split a bytearray into groups
-static int split_bytearray_dtype16(u_int8_t *src, Py_ssize_t len, u_int8_t **buffers,
-                           int bits_mode, int bytes_mode, int is_review,
-                           int threads) {
-  if (bits_mode == 1) {  // reoreder exponent
-    reorder_all_floats_dtype16(src, len);
-  }
-
-  Py_ssize_t half_len = len / 2;
-  switch (bytes_mode) {
-  case 10:  // 2b01_010 - Byte Group to two different groups
-    buffers[0] = PyMem_Malloc(half_len);
-    buffers[1] = PyMem_Malloc(half_len);
-
-    if (buffers[0] == NULL || buffers[1] == NULL) {
-      PyMem_Free(buffers[0]);
-      PyMem_Free(buffers[1]);
-      return -1;
-    }
-
-    u_int8_t *dst0 = buffers[0];
-    u_int8_t *dst1 = buffers[1];
-
-    for (Py_ssize_t i = 0; i < len; i += 2) {
-      *dst0++ = src[i];
-      *dst1++ = src[i + 1];
-    }
-    break;
-
-  case 8:  // 4b1000 - Truncate MSByte
-           // We are refering to the MSBbyte as little endian, thus we omit buf2
-  case 1:  // 4b1000 - Truncate LSByte
-    // We are refering to the LSByte  as a little endian, thus we omit buf1
-    buffers[0] = PyMem_Malloc(half_len);
-    buffers[1] = NULL;
-
-    if (buffers[0] == NULL) {
-      PyMem_Free(buffers[0]);
-      return -1;
-    }
-
-    dst0 = buffers[0];
-
-    if (bytes_mode == 1) {
-      for (Py_ssize_t i = 0; i < len; i += 2) {
-        *dst0++ = src[i];
-      }
-    } else {
-      for (Py_ssize_t i = 0; i < len; i += 2) {
-        *dst0++ = src[i + 1];
-      }
-    }
-    break;
-
-  default:
-    // we are not support this splitting bytes_mode
-    return -1;
-  }
-  return 0;
-}
-
-///////////////////////////////////
-/////////  Combine Functions //////
-///////////////////////////////////
-
-// Reordering function for float bits
-static uint32_t revert_float_bits_dtype16(float number) {
-  union {
-    float f;
-    uint32_t u;
-  } value = {.f = number};
-
-  uint32_t sign = (value.u << 8) & 0x80008000;
-  uint32_t exponent = (value.u >> 1) & 0x7F807F80;
-  uint32_t mantissa = (value.u) & 0x7F007F;
-  return sign | exponent | mantissa;
-}
-
-// Helper function to reorder all floats in a bytearray
-static void revert_all_floats_dtype16(u_int8_t *src, Py_ssize_t len) {
-  uint32_t *uint_array = (uint32_t *)src;
-  Py_ssize_t num_floats = len / sizeof(uint32_t);
-  for (Py_ssize_t i = 0; i < num_floats; i++) {
-    uint_array[i] = revert_float_bits_dtype16(*(float *)&uint_array[i]);
-  }
-}
-
-// Helper function to combine four buffers into a single bytearray
-static int combine_buffers_dtype16(u_int8_t *buf1, u_int8_t *buf2, u_int8_t *combinePtr,
-                           Py_ssize_t half_len, int bits_mode, int bytes_mode,
-                           int threads) {
-  Py_ssize_t total_len = half_len * 2;
-
-  u_int8_t *dst;
-  dst = combinePtr;
-
-  switch (bytes_mode) {
-  case 10: // 2b01_010 - Byte Group to two different groups
-    for (Py_ssize_t i = 0; i < half_len; i++) {
-      *dst++ = buf1[i];
-      *dst++ = buf2[i];
-    }
-    break;
-
-  case 8: // 4b1000 - Truncate MSByte
-          // We are refering to the MSByte as a little endian, thus we omit buf2
-  case 1: // 4b001 - Truncate LSByte
-          // We are refering to the LSByte as a little endian, thus we omit buf1
-
-    if (bytes_mode == 8) {
-      for (Py_ssize_t i = 0; i < half_len; i++) {
-        *dst++ = 0;
-        *dst++ = buf1[i];
-      }
-    } else {
-      for (Py_ssize_t i = 0; i < half_len; i++) {
-        *dst++ = buf1[i];
-        *dst++ = 0;
-      }
-    }
-    break;
-
-  default:
-    // we are not supporting this splitting bytes_mode
-    return -1;
-  }
-  // printf("combinePtr %zu\n ",combinePtr);
-  // printf("dst %zu\n ", dst);
-  //  Revert the reordering of all floats if needed
-  if (bits_mode == 1) {
-    revert_all_floats_dtype16(combinePtr, total_len);
-  }
-  return 0;
-}
-
-///////////// helper function to prepare the split data
-//////////////////////////////////////
+//////////////////////// Help function /////////////
 u_int8_t *prepare_split_results(size_t header_len, size_t numBuf,
                                 size_t numChunks, u_int8_t *header,
                                 u_int8_t *compressedData[numBuf][numChunks],
@@ -209,7 +48,7 @@ u_int8_t *prepare_split_results(size_t header_len, size_t numBuf,
   memcpy(resultBuf + offset, cumulativeChunksSize, cumulativeChunksSizeLen);
   offset += cumulativeChunksSizeLen;
 
-  for (uint32_t b = 0; b < numBuf; b++) {
+  for (int b = 0; b < numBuf; b++) {
     for (uint32_t c = 0; c < numChunks; c++) {
       memcpy(resultBuf + offset, compressedData[b][c], compChunksSize[b][c]);
       // printf("\n\n resultBuf + offset %zu\n\n\n ", resultBuf + offset);
@@ -297,7 +136,7 @@ PyObject *py_split_dtype(PyObject *self, PyObject *args) {
     //    }
     // Compression on each Buf
 
-    for (uint32_t b = 0; b < numBuf; b++) {
+    for (int b = 0; b < numBuf; b++) {
       if (isThCheck[b] == 0 &&
           curChunk >=
               checkCompTh) {  // check that we really need to compress this buf
@@ -311,7 +150,7 @@ PyObject *py_split_dtype(PyObject *self, PyObject *args) {
       compressedData[b][curChunk] = PyMem_Malloc(bgChunkSize);
       if (!compressedData[b][curChunk]) {
         PyErr_SetString(PyExc_MemoryError, "Failed to allocate memory");
-        for (uint32_t j = 0; j < numBuf; j++) {
+        for (int j = 0; j < numBuf; j++) {
           for (uint32_t c = 0; c < curChunk - 1; c++) {
             PyMem_Free(compressedData[j][c]);
           }
@@ -372,7 +211,7 @@ PyObject *py_split_dtype(PyObject *self, PyObject *args) {
 
   // Freeing compressedData array
   for (uint32_t c = 0; c < numChunks; c++) {
-    for (uint32_t b = 0; b < numBuf; b++) {
+    for (int b = 0; b < numBuf; b++) {
       if (buffers[c][b] != NULL) {
         PyMem_Free(buffers[c][b]);
       }
@@ -380,7 +219,7 @@ PyObject *py_split_dtype(PyObject *self, PyObject *args) {
   }
 
   for (uint32_t c = 0; c < numChunks; c++) {
-    for (uint32_t b = 0; b < numBuf; b++) {
+    for (int b = 0; b < numBuf; b++) {
       if (compChunksType[b][c] == 1) {
         PyMem_Free(compressedData[b][c]);
       }
@@ -425,7 +264,7 @@ PyObject *py_combine_dtype(PyObject *self, PyObject *args) {
   compCumulativeChunksPos[0][0] = 0;
   compCumulativeChunksPos[1][0] = 0;
 
-  for (uint32_t b = 0; b < numBuf; b++) {
+  for (int b = 0; b < numBuf; b++) {
     for (uint32_t c = 0; c < numChunks; c++) {
       compChunksType[b][c] = (*ptrChunksType++);
       cumulativeChunksSize[b][c] = (*ptrChunksCumulative++);
@@ -433,7 +272,7 @@ PyObject *py_combine_dtype(PyObject *self, PyObject *args) {
   }
 
   for (uint32_t c = 0; c < numChunks; c++) {
-    for (uint32_t b = 0; b < numBuf; b++) {
+    for (int b = 0; b < numBuf; b++) {
       compCumulativeChunksPos[b][c + 1] = cumulativeChunksSize[b][c];
       CompChunksLen[b][c] =
           compCumulativeChunksPos[b][c + 1] - compCumulativeChunksPos[b][c];
@@ -441,7 +280,7 @@ PyObject *py_combine_dtype(PyObject *self, PyObject *args) {
   }
 
   for (size_t c = 0; c < numChunks; c++) {
-    for (uint32_t b = 0; b < numBuf; b++) {
+    for (int b = 0; b < numBuf; b++) {
       if (compChunksType[b][c] == 0) { // no compression is needed
       } else {
         if (compChunksType[b][c] == 1) { // open with Huffman compression
@@ -524,7 +363,7 @@ PyObject *py_combine_dtype(PyObject *self, PyObject *args) {
       PyByteArray_FromStringAndSize((const char *)resultBuf, origSize);
 
   for (size_t c = 0; c < numChunks; c++) {
-    for (uint32_t b = 0; b < numBuf; b++) {
+    for (int b = 0; b < numBuf; b++) {
       if (compChunksType[b][c] > 0) {
         free(deCompressedData[b][c]);
       }
