+++ setup.py
@@ -20,6 +20,7 @@ split_dtype_extension = Extension(
         "csrc/split_dtype_module.c",
         "csrc/split_dtype32.c",
         "csrc/split_dtype.c",
+        "csrc/data_manipulation_dtype16.c",
         "include/FiniteStateEntropy/lib/fse_compress.c",
         "include/FiniteStateEntropy/lib/fse_decompress.c",
         "include/FiniteStateEntropy/lib/huf_compress.c",
